pkt_size is std::size_t

+++ include/spead2/recv_tcp.h
@@ -57,7 +57,7 @@ private:
     /// The tail of the buffer, after this there is no more data
     std::uint8_t *tail;
     /// Size of the current packet being parsed. 0 means no packet is being parsed
-    std::uint64_t pkt_size = 0;
+    std::size_t pkt_size = 0;
     /// Buffer size to set on peer
     std::size_t buffer_size;
     /// Number of bytes that need to be skipped (used when pkt_size > max_size)
